build-dm-verity-task: remove the PODVM_PAYLOAD_IMAGE parameter
Having a parameter on the task is causing problems during our builds.
Conforma (i.e: "Enterprise Contract") checks that the parameter is valid for the
task. Now whenever the podvm-payload image is rebuilt, this check fails until
we update the dm-verity image with the new reference.

This happens on the PR that modifies podvm-payload (which needs to be merged
despite the failure), but it also happen after the PR is merged and podvm-payload
is rebuilt, for EVERY pull request on all our repositories, until the dm-verity
image is rebuilt - which can take some time.

This commit tries to work around that issue by putting the reference to the
podvm-payload image in the "init" Dockerfile used by the task. The Dockerfile
now takes the binaries from the podvm-payload image at build time, and we
copy them to the VM directly from the container, without a need to give the task
a reference.

The reference to podvm-payload in the Dockerfile will be updated as usual using nudges.

Signed-off-by: Julien Ropé <jrope@redhat.com>