dm-verity: enable image signature verification
The kata-agent version in OSC 1.10 is in charge of the image signature
verification workflow. To have image signature enabled, the agent should
be configured with:
 * `enable_signature_verification=true`
 * the KBS URI (`image_policy_file`) of the containers-policy.json

In peer pods, the kata-agent configuration is mounted at
/run/peerpod/agent-config.toml by the process-user-data (at podvm launch time)
but that file cannot be attested, hence a malicious host can tamper the
configuration (e.g. to turn off the signature). A workaround is to have the agent
configuration statically in the podvm image, which is attested.

This will create a custom agent-config.toml in the podvm image. And
adjust the kata-agent service to use that file instead of
/run/peerpod/agent-config.toml.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>